Move proxy impl to a different file
