+++ catowseriOS/CottonViewModels/TabPreviewsViewModel/TabsPreviewState.swift
@@ -14,7 +14,7 @@ public enum TabsPreviewState<C: TabsPreviewsStateContext>: ViewModelState {
     /// Maybe it is not needed state, but it is required for scalability when some user will have 100 tabs
     case loading
     /// Actual collection for tabs, at least one tab always will be in it
-    case tabs(dataSource: [CoreBrowser.Tab], selectedId: UUID?)
+    case tabs([CoreBrowser.Tab], CoreBrowser.Tab.ID?)
 
     /// Number of items for each state
     public var itemsNumber: Int {
@@ -43,35 +43,35 @@ public enum TabsPreviewState<C: TabsPreviewsStateContext>: ViewModelState {
         case .load:
             if let info = await context?.load() {
                 nextState = .tabs(
-                    dataSource: info.tabs,
-                    selectedId: info.selectedTabUUID
+                    info.tabs,
+                    info.selectedTabUUID
                 )
             } else {
                 throw TabsPreviewsError.failToLoad
             }
         case .closeTab(index: let index):
-            guard case let .tabs(tabs, selectedId) = self else {
+            guard case let .tabs(tabs, _) = self else {
                 throw TabsPreviewsError.tabsNotLoadedToClose
             }
             guard let info = try await context?.close(at: index, from: tabs) else {
                 throw TabsPreviewsError.nilStateContext
             }
             nextState = .tabs(
-                dataSource: info.tabs,
-                selectedId: info.selectedTabUUID
+                info.tabs,
+                info.selectedTabUUID
             )
         case .select(let tab):
             guard case let .tabs(tabs, _) = self else {
                 throw TabsPreviewsError.tabsNotLoadedToClose
             }
             try await context?.select(tab)
-            /// Set new selected id
-            nextState = .tabs(dataSource: tabs, selectedId: tab.id)
+            // Set new selected id
+            nextState = .tabs(tabs, tab.id)
         case .addDefaultTab:
             if let info = try await context?.addDefaultTab() {
                 nextState = .tabs(
-                    dataSource: info.tabs,
-                    selectedId: info.selectedTabUUID
+                    info.tabs,
+                    info.selectedTabUUID
                 )
             } else {
                 throw TabsPreviewsError.nilStateContext
@@ -79,8 +79,8 @@ public enum TabsPreviewState<C: TabsPreviewsStateContext>: ViewModelState {
         case let .addTab(tab: tab, index: index):
             if let info = try await context?.addTab(tab, at: index) {
                 nextState = .tabs(
-                    dataSource: info.tabs,
-                    selectedId: info.selectedTabUUID
+                    info.tabs,
+                    info.selectedTabUUID
                 )
             } else {
                 throw TabsPreviewsError.nilStateContext
@@ -94,16 +94,16 @@ public enum TabsPreviewState<C: TabsPreviewsStateContext>: ViewModelState {
         with context: Context?,
         onComplete: @escaping (Result<BaseState, Error>) -> Void
     ) {
+        guard let context else {
+            onComplete(.failure(TabsPreviewsError.nilStateContext))
+            return
+        }
         switch action {
         case .load:
-            guard let context else {
-                onComplete(.failure(TabsPreviewsError.nilStateContext))
-                return
-            }
             context.load(onComplete: { info in
                 let nextState: TabsPreviewState = .tabs(
-                    dataSource: info.tabs,
-                    selectedId: info.selectedTabUUID
+                    info.tabs,
+                    info.selectedTabUUID
                 )
                 onComplete(.success(nextState))
             })
@@ -112,16 +112,12 @@ public enum TabsPreviewState<C: TabsPreviewsStateContext>: ViewModelState {
                 onComplete(.failure(TabsPreviewsError.tabsNotLoadedToClose))
                 return
             }
-            guard let context else {
-                onComplete(.failure(TabsPreviewsError.nilStateContext))
-                return
-            }
             context.close(at: index, from: tabs) { result in
                 switch result {
                 case .success(let info):
                     let nextState: TabsPreviewState = .tabs(
-                        dataSource: info.tabs,
-                        selectedId: info.selectedTabUUID
+                        info.tabs,
+                        info.selectedTabUUID
                     )
                     onComplete(.success(nextState))
                 case .failure(let error):
@@ -129,10 +125,6 @@ public enum TabsPreviewState<C: TabsPreviewsStateContext>: ViewModelState {
                 }
             }
         case .select(let tab):
-            guard let context else {
-                onComplete(.failure(TabsPreviewsError.nilStateContext))
-                return
-            }
             context.select(tab) { result in
                 switch result {
                 case .success:
@@ -141,8 +133,8 @@ public enum TabsPreviewState<C: TabsPreviewsStateContext>: ViewModelState {
                         return
                     }
                     let nextState: TabsPreviewState = .tabs(
-                        dataSource: tabs,
-                        selectedId: tab.id
+                        tabs,
+                        tab.id
                     )
                     onComplete(.success(nextState))
                 case .failure(let error):
@@ -150,9 +142,32 @@ public enum TabsPreviewState<C: TabsPreviewsStateContext>: ViewModelState {
                 }
             }
         case .addDefaultTab:
-            onComplete(.failure(TabsPreviewsError.notImplementedYet))
+            Task {
+                do {
+                    let info = try await context.addDefaultTab()
+                    let nextState: TabsPreviewState = .tabs(
+                        info.tabs,
+                        info.selectedTabUUID
+                    )
+                    onComplete(.success(nextState))
+                } catch {
+                    onComplete(.failure(error))
+                }
+            }
         case let .addTab(tab: tab, index: index):
-            onComplete(.failure(TabsPreviewsError.notImplementedYet))
+            Task {
+                do {
+                    // This won't call any use cases or persistence storage
+                    let info = try await context.addTab(tab, at: index)
+                    let nextState: TabsPreviewState = .tabs(
+                        info.tabs,
+                        info.selectedTabUUID
+                    )
+                    onComplete(.success(nextState))
+                } catch {
+                    onComplete(.failure(error))
+                }
+            }
         }
     }
