+++ catowseriOS/CottonViewModels/TabPreviewsViewModel/TabsPreviewsAction.swift
@@ -11,11 +11,15 @@ import ViewModelKit
 
 /// Tab previews view model action
 public enum TabsPreviewsAction: ViewModelAction {
+    /// Load all the tabs
     case load
+    /// Close specific tab at index
     case closeTab(index: Int)
+    /// Select different tab
     case select(CoreBrowser.Tab)
+    /// Handle user tap on + tab button
     case addDefaultTab
-    #warning("TODO: remove when observation moves to view model")
+    /// Add custom tab to in-memory state/cache only
     case addTab(tab: CoreBrowser.Tab, index: Int)
     
     public static let allCases: [TabsPreviewsAction] = [
